+++ frontend/src/metabase/visualizations/components/ScalarValue/utils.unit.spec.ts
@@ -1,5 +1,5 @@
 import * as measureText from "metabase/lib/measure-text";
-import { FontStyle } from "metabase/visualizations/types/measure-text";
+import { FontStyle } from "metabase/visualizations/shared/types/measure-text";
 import { findSize } from "./utils";
 
 jest.doMock("metabase/lib/measure-text", () => ({
