+++ frontend/src/metabase/visualizations/lib/settings/goal.ts
@@ -1,6 +1,6 @@
 import { t } from "ttag";
 import { VisualizationSettings } from "metabase-types/api";
-import { ChartGoal } from "metabase/visualizations/types/settings";
+import { ChartGoal } from "metabase/visualizations/shared/types/settings";
 import { getStackingOffset } from "./stacking";
 
 const getGoalValue = (value: number, isPercent: boolean) =>
