+++ frontend/src/metabase/visualizations/shared/components/RowChart/utils/layout.ts
@@ -4,10 +4,10 @@ import { stack, stackOffsetExpand, stackOffsetNone } from "d3-shape";
 import type { Series as D3Series, SeriesPoint } from "d3-shape";
 import { scaleBand } from "@visx/scale";
 import type { ScaleBand, ScaleLinear } from "d3-scale";
-import { TextMeasurer } from "metabase/visualizations/types/measure-text";
-import { ChartFont } from "metabase/visualizations/types/theme";
-import { Margin } from "metabase/visualizations/types/layout";
-import { ContinuousScaleType } from "metabase/visualizations/types/scale";
+import { TextMeasurer } from "metabase/visualizations/shared/types/measure-text";
+import { Margin } from "metabase/visualizations/shared/types/layout";
+import { ContinuousScaleType } from "metabase/visualizations/shared/types/scale";
+import { ChartFont } from "metabase/visualizations/shared/types/style";
 import { LABEL_PADDING } from "../constants";
 import { Series } from "../types";
 import { createStackedXScale, createXScale, createYScale } from "./scale";
