+++ frontend/src/metabase/visualizations/shared/types/scale.ts
