+++ frontend/src/metabase/visualizations/shared/types/settings.ts
