+++ frontend/src/metabase/visualizations/shared/utils/colors.ts
@@ -1,6 +1,6 @@
 import { VisualizationSettings } from "metabase-types/api";
 import { getColorsForValues } from "metabase/lib/colors/charts";
-import { Series } from "../types";
+import { Series } from "../components/RowChart/types";
 
 export const getSeriesColors = <TDatum, TSeriesInfo>(
   settings: VisualizationSettings,
