--- frontend/src/metabase/visualizations/types/data.ts
@@ -1,20 +1,0 @@
-import { DatasetColumn, RowValue } from "metabase-types/api";
-
-export type SeriesInfo = {
-  metricColumn: DatasetColumn;
-  dimensionColumn: DatasetColumn;
-  breakoutValue?: RowValue;
-};
-
-export type MetricValue = number | null;
-export type MetricName = string;
-export type BreakoutName = string;
-export type MetricDatum = { [key: MetricName]: MetricValue };
-
-export type GroupedDatum = {
-  dimensionValue: RowValue;
-  metrics: MetricDatum;
-  breakout?: { [key: BreakoutName]: MetricDatum };
-};
-
-export type GroupedDataset = GroupedDatum[];
